incorporated PR feedback to merge resources in a single module spring.py
