+++ tools/c7n_azure/c7n_azure/resources/spring.py
@@ -4,6 +4,7 @@
 from c7n_azure.provider import resources
 from c7n_azure.query import ChildResourceManager, ChildTypeInfo
 from c7n_azure.utils import ResourceIdParser
+from c7n_azure.resources.arm import ArmResourceManager
 
 
 @resources.register('spring-app')
@@ -41,3 +42,33 @@ class SpringApp(ChildResourceManager):
             return {'resource_group_name': 
                     ResourceIdParser.get_resource_group(parent_resource['id']),
                     'service_name': parent_resource['name']}
+
+
+@resources.register('spring-service-instance')
+class SpringServiceInstance(ArmResourceManager):
+    """Azure Spring Service Instance Resource
+
+    :example:
+
+    Returns Spring Service Instance resources
+
+    .. code-block:: yaml
+
+         policies:
+          - name: basic-spring-service-instance
+            resource: azure.spring-service-instance
+
+    """
+
+    class resource_type(ArmResourceManager.resource_type):
+        doc_groups = ['Compute']
+
+        service = 'azure.mgmt.appplatform'
+        client = 'AppPlatformManagementClient'
+        enum_spec = ('services', 'list_by_subscription', None)
+        default_report_fields = (
+            'name',
+            'location',
+            'resourceGroup'
+        )
+        resource_type = 'Microsoft.AppPlatform/Spring'
